Do some work refactoring Header and Footer, moving parts out into seperate components to keep the files clean. Also removed the cards as I want to completely change how they look and are done
