+++ src/lib/components/footer/footercredit.svelte
@@ -1,0 +1,26 @@
+<p class="footertext">
+	Made by
+	<a href="https://github.com/johnnyg-git" target="_blank" rel="noopener noreferrer">
+		Joseph Goold (Johnny Johnny)
+	</a>
+</p>
+
+<style>
+	.footertext {
+		color: var(--text-color-less);
+		font-size: 0.8em;
+		margin: 0;
+	}
+
+	a {
+		color: var(--primary-color);
+		text-decoration: none;
+		font-weight: normal;
+		transition: color 0.3s ease;
+		will-change: color;
+	}
+
+	a:hover {
+		color: var(--accent-color);
+	}
+</style>
