+++ src/routes/+layout.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import '../app.css';
 
-	import Header from '$lib/components/Header.svelte';
-	import Footer from '$lib/components/Footer.svelte';
+	import Header from '$lib/components/header/header.svelte';
+	import Footer from '$lib/components/footer/footer.svelte';
 
 	let { children } = $props();
 </script>
@@ -17,9 +17,6 @@
 
 <style>
 	.container {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
 		width: 100%;
 		min-height: 100vh;
 		min-height: 100vdh;
